requirements.txt
streamlit
elevenlabs
python-dotenv
spotipy